Move VDO's to vdo.rs
